I moved the structure to a Class
